+++ CHANGES.md
@@ -1,8 +1,14 @@
 Contargo Types
 ===============
+
+## v0.17.1
+* use PhoneNumber in ContactInformation instead of String for phone- and mobile numbers
+* change PhoneNumber only format number if method is called. Has country code, raw phone number and extension 
+to a phone number. 
+
 ## v0.17.0
 
-* Adds PhoneNumber to formate phone numbers and get information to a spezific phone number, 
+* Adds PhoneNumber to format phone numbers and get information to a specific phone number, 
 like the country code or country calling code.
 * Change the amount of characters for license plates to max. 15 characters.